[JBRes-10600] Move Dockerfile text analysis into common-utils
