+++ backend-utils/pyproject.toml
@@ -17,6 +17,7 @@ dependencies = [
     "google-cloud-secret-manager>=2.24.0",
     "google-cloud-storage>=3.13.0",
     "idegym-api",
+    "idegym-common-utils",
     "jinja2>=3.1.6",
     "kubernetes-asyncio>=32.3.2",
     "opentelemetry-api>=1.44.0",
